+++ sageintacctsdk/apis/constants.py
@@ -1,0 +1,192 @@
+dimensions_fields_mapping = {
+    'GLACCOUNT': [
+        'RECORDNO',
+        'ACCOUNTNO',
+        'TITLE',
+        'ACCOUNTTYPE',
+        'NORMALBALANCE',
+        'CLOSINGTYPE',
+        'STATUS',
+        'CATEGORY',
+        'ALTERNATIVEACCOUNT'
+    ],
+    'CHECKINGACCOUNT': [
+        'BANKACCOUNTID',
+        'BANKACCOUNTNO',
+        'GLACCOUNTNO',
+        'BANKNAME',
+        'ROUTINGNO',
+        'BRANCHID',
+        'BANKACCOUNTTYPE',
+        'DEPARTMENTID',
+        'LOCATIONID',
+        'STATUS',
+        'RECORDNO',
+        'ACHBANKID',
+        'COMPANYNAME'
+    ],
+    'CONTACT': [
+        'RECORDNO',
+        'CONTACTNAME',
+        'COMPANYNAME',
+        'FIRSTNAME',
+        'LASTNAME',
+        'INITIAL',
+        'PRINTAS',
+        'TAXABLE',
+        'MAILADDRESS.ADDRESS1'
+    ],
+    'DEPARTMENT': [
+        'RECORDNO',
+        'DEPARTMENTID',
+        'TITLE',
+        'PARENTKEY',
+        'PARENTID',
+        'SUPERVISORNAME',
+        'STATUS',
+        'CUSTTITLE'
+    ],
+    'EXPENSEPAYMENTTYPE': [
+        'RECORDNO',
+        'NAME',
+        'DESCRIPTION',
+        'NONREIMBURSABLE',
+        'OFFSETACCTNO',
+        'OFFSETACCTTITLE',
+        'STATUS',
+        'WHENCREATED',
+        'WHENMODIFIED',
+        'CREATEDBY',
+        'MODIFIEDBY',
+        'MEGAENTITYKEY',
+        'MEGAENTITYID',
+        'MEGAENTITYNAME'
+    ],
+    'EEXPENSES': [
+        'RECORDNO',
+        'RECORDID',
+        'WHENCREATED',
+        'WHENPOSTED',
+        'TOTALENTERED',
+        'STATE',
+        'TOTALDUE',
+        'DESCRIPTION',
+        'CURRENCY',
+        'BASECURR',
+        'MEMO'
+    ],
+    'EEACCOUNTLABEL': [
+        'RECORDNO',
+        'ACCOUNTLABEL',
+        'DESCRIPTION',
+        'GLACCOUNTNO',
+        'GLACCOUNTTITLE',
+        'STATUS',
+        'ITEMID'
+    ],
+    'ITEM': [
+        'RECORDNO',
+        'ITEMID',
+        'STATUS',
+        'MRR',
+        'NAME',
+        'EXTENDED_DESCRIPTION',
+        'PRODUCTLINEID',
+        'GLGROUP',
+        'ITEMTYPE'
+    ],
+    'LOCATION': [
+        'RECORDNO',
+        'LOCATIONID',
+        'NAME',
+        'PARENTID',
+        'STATUS',
+        'CURRENCY'
+    ],
+    'PROJECT': [
+        'RECORDNO',
+        'PROJECTID',
+        'NAME',
+        'DESCRIPTION',
+        'CURRENCY',
+        'PROJECTCATEGORY',
+        'PROJECTSTATUS',
+        'PARENTKEY',
+        'PARENTID',
+        'PARENTNAME',
+        'STATUS',
+        'CUSTOMERKEY',
+        'CUSTOMERID',
+        'CUSTOMERNAME',
+        'PROJECTTYPE',
+        'DEPARTMENTNAME',
+        'LOCATIONID',
+        'LOCATIONNAME',
+        'BUDGETID',
+        'MEGAENTITYID',
+        'MEGAENTITYNAME'
+    ],
+    'TASK': [
+        'RECORDNO',
+        'TASKID',
+        'PARENTKEY',
+        'PARENTID',
+        'NAME',
+        'PARENTTASKNAME',
+        'PROJECTKEY',
+        'PROJECTID',
+        'PROJECTNAME',
+        'ITEMKEY',
+        'ITEMID',
+        'ITEMNAME',
+        'DESCRIPTION',
+        'BILLABLE',
+        'TASKNO',
+        'TASKSTATUS',
+        'CLASSID',
+        'CLASSNAME',
+        'CLASSKEY',
+        'ROOTPARENTKEY',
+        'ROOTPARENTID',
+        'ROOTPARENTNAME'
+    ],
+    'VENDOR': [
+        'RECORDNO',
+        'NAME',
+        'VENDORID',
+        'PARENTKEY',
+        'PARENTID',
+        'PARENTNAME',
+        'DISPLAYCONTACT.CONTACTNAME',
+        'DISPLAYCONTACT.COMPANYNAME',
+        'DISPLAYCONTACT.FIRSTNAME',
+        'DISPLAYCONTACT.LASTNAME',
+        'DISPLAYCONTACT.INITIAL',
+        'DISPLAYCONTACT.PRINTAS',
+        'DISPLAYCONTACT.PHONE1',
+        'DISPLAYCONTACT.PHONE2',
+        'DISPLAYCONTACT.EMAIL1',
+        'DISPLAYCONTACT.EMAIL2',
+        'VENDORACCOUNTNO',
+        'VENDTYPE',
+        'ACCOUNTLABEL',
+        'APACCOUNT',
+        'APACCOUNTTITLE',
+        'STATUS'
+    ],
+    'CREDITCARD': [
+        'RECORDNO',
+        'CARDID',
+        'DESCRIPTION',
+        'CARDTYPE',
+        'EXP_MONTH',
+        'EXP_YEAR',
+        'COMMCARD',
+        'STATUS',
+        'VENDORID',
+        'DEPT',
+        'LOCATION',
+        'LIABILITYTYPE',
+        'OUTSOURCECARD'
+    ]
+}
